refactor: rename gee/ndvi.py to gee/sat_index_info.py
